+++ plugins/http_file_uploader/lib/src/file_handler.dart
@@ -4,7 +4,7 @@ import 'package:http_file_uploader/http_file_uploader.dart';
 import 'package:http_file_uploader/src/http_ext.dart';
 
 /// [HttpFileHandler] handle the file upload using the [http.Client]
-class HttpFileHandler extends SocketFileHandler {
+class HttpFileHandler extends SocketFileHandler<http.Response> {
   /// [client] used to upload the file
   ///
   /// [path], [method], [headers], [body] are [http.Client.send] parameters
@@ -35,6 +35,6 @@ class HttpFileHandler extends SocketFileHandler {
           fileKey: fileKey,
           onProgress: onProgress,
         )
-        .then((value) {});
+        .then(fileParser);
   }
 }