+++ tools/file_uploader_socket_interfaces/lib/src/default.dart
@@ -1,0 +1,17 @@
+/// default file upload method
+const kFileUploadMethod = 'POST';
+
+/// default file key
+const kFileKey = 'file';
+
+/// default present method
+const kPresentMethod = 'POST';
+
+/// default file chunk upload method
+const kChunkMethod = 'POST';
+
+/// default status method
+const kStatusMethod = 'HEAD';
+
+/// default chunk parser
+void kChunkParser(dynamic _) {}