+++ tools/file_uploader_socket_interfaces/lib/src/interfaces/file_interfaces.dart
@@ -1,0 +1,4 @@
+/// callback to validate chunk upload
+typedef ChunkParser<T> = void Function(
+  T response,
+);