+++ tools/file_uploader_socket_interfaces/lib/src/interfaces/interfaces.dart
@@ -1,2 +1,3 @@
 export './chunked_file_interfaces.dart';
+export './file_interfaces.dart';
 export './restorable_chunked_file_interfaces.dart';