+++ tools/file_uploader_socket_interfaces/CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.2.2] - 2024-06-22
+
+### Fix
+
+- Fixed chunk parser
+
+## [1.2.1] - 2024-06-22
+
+### Fix
+
+- Fixed chunk parser
+
+## [1.2.0] - 2024-06-22
+
+### Added
+
+- Added chunk parser
+
 ## [1.1.0] - 2024-06-22
 
 ### Added